Fixing the ChatGPT mistake omitting an API Key
ME: Doesn't this need an API key to make an API call to chatGPT that is different from the accesstoken for webex teams?

ChatGPT: Yes, you are correct. In order to make API calls to the ChatGPT API, you will need an API key. The access token that you use for the Webex Teams API is not the same as the API key that is needed to access the ChatGPT API. You can sign up for an API key on the OpenAI website and then use it in the code.