fixed another bug from the initial generated code
ME: The line "const { choice } = response.data;"    doesn't appear to be the correct method of retrieving the API response for choice.
You're correct, I apologize for the mistake in my previous response. The code I provided for extracting the "choice" value from the API response is not correct.

ChatGPT: The correct way to extract the "choice" value from the API response is by using the dot notation to access the property of the object. Here is the corrected version of the code snippet: